Cleanup only for the Try box

+++ scripts/minimize.sh
@@ -11,19 +11,19 @@ echo "==> Removing all linux kernels except the currrent one"
 dpkg --list | awk '{ print $2 }' | grep 'linux-image-3.*-generic' | grep -v $(uname -r) | xargs apt-get -y purge
 echo "==> Removing linux source"
 dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
-echo "==> Removing development packages"
-dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
-echo "==> Removing documentation"
-dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
-echo "==> Removing development tools"
+if [ "$BOX_TYPE" = "Try" ]; then
+	echo "==> Removing development packages"
+	dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
+	echo "==> Removing documentation"
+	dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
+#echo "==> Removing development tools"
 #dpkg --list | grep -i compiler | awk '{ print $2 }' | xargs apt-get -y purge
 #apt-get -y purge cpp gcc g++
 #apt-get -y purge build-essential git
-echo "==> Removing default system Ruby"
-apt-get -y purge ruby ri doc
-echo "==> Removing default system Python"
-apt-get -y purge build-essential python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl libffi5
-if [ "$BOX_TYPE" = "Try" ]; then
+	echo "==> Removing default system Ruby"
+	apt-get -y purge ruby ri doc
+	echo "==> Removing default system Python"
+	apt-get -y purge build-essential python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl libffi5
 	echo "==> Removing fluff"
 	apt-get -y purge byobu git git-man htop multitail php5-xdebug php5-xhprof postfix subversion tmux vim-common
 
@@ -49,8 +49,10 @@ while [ -n "$(deborphan --guess-all --libdevel)" ]; do
 done
 apt-get -y purge deborphan dialog
 
-echo "==> Removing man pages"
-rm -rf /usr/share/man/*
+if [ "$BOX_TYPE" = "Try" ]; then
+	echo "==> Removing man pages"
+	rm -rf /usr/share/man/*
+fi
 echo "==> Removing APT files"
 find /var/lib/apt -type f | xargs rm -f
 echo "==> Removing any docs"
